+++ src/me/mafrans/loggbok/Volume.kt
@@ -1,7 +1,7 @@
 package me.mafrans.loggbok
 
-class Volume(index: UShort, book: Book) {
-    var index: UShort = index;
+class Volume(index: Int, book: Book) {
+    var index: Int = index;
     var pages: Set<Page> = HashSet();
     val book: Book = book;
 }
